renamed model.py to schema.py
